Experimented with putAttachment to reduce disk use
Really confusing. The raw leveldb is ~200 kb. The JSON printout of the
leveldb is 520 kb. But the SQLite PouchDB file is 5 MB? And if I use
putAttachment, it's 11 MB? (Presumably the latter because it's making
twice as many revs: one for the doc update and one for the attachment.)
Using the Level backend (instead of SQLite) is ~5 MB too, so it's not
that. For ALL these sizes, I'm compacting the database before checking.

Even when I save allDocs' output (including attachments) to JSON, the
result is just ~1 MB. What gives?